rename 'queries' to 'from_table' and move under 'linters' module
