+++ tests/unit/source_code/linters/test_pyspark.py
@@ -5,7 +5,7 @@ from astroid import Call, Const, Expr  # type: ignore
 from databricks.labs.ucx.source_code.base import Deprecation, CurrentSessionState
 from databricks.labs.ucx.source_code.python.python_ast import Tree, TreeHelper
 from databricks.labs.ucx.source_code.linters.pyspark import TableNameMatcher, SparkSqlPyLinter
-from databricks.labs.ucx.source_code.queries import FromTableSqlLinter
+from databricks.labs.ucx.source_code.linters.from_table import FromTableSqlLinter
 
 
 def test_spark_no_sql(empty_index):
